requirements.txt
openai>=1.0.0
chromadb>=0.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
loguru>=0.7.0
